+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/domain/model/AroundPlace.kt
@@ -6,5 +6,5 @@ data class AroundPlace (
     val disability: List<String>,
     val image: String,
     val name: String,
-    val placeId: Int
+    val placeId: Long
 )
